Move code out of shared/_doc_utils/__init__.py
